fix(data): fix serializer problems hit when serializer = false
In an Action, `serializer = false` is something we're allowed to do.
However, if we're doing a POST of data into Denali (using a webhook)
and have a resulting request.body, we weren't checking whether a
serializer was desired, but were instead using the application
serializer on the body.

In my case, I wanted the raw POST body to make it through untouched.
This fixes that ...